Lines (25 - 27): found the difference between  the predicted and observed temperature (error).
Lines (30 - 32): Calculated the mae error for both uncalibrated and hybrid models.

Lines (35 - 36) : Calculated the rmse error for both uncalibrated and hybrid models.

lines (39 - 43): Calculated the mare error for both uncalibrated and hybrid models.

lines (45 - 54) Calculated the mare error for both uncalibrated and hybrid models. In this case the observed temperature values that equal to 0 were turned to a very small value.

Lines (56 - 63) Calculated the mare error for both uncalibrated and hybrid models. In this case the we removed all of the `-inf, inf`  values.

Lines (65 - 67): we combined the calculated metrics into one table for comparison.

Lines (70 - 87): plotting